requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.8.2
dj-database-url==0.2.2
dj-static==0.0.5
django-bootstrap-toolkit==2.15.0
django-crispy-forms==1.3.2
django-markdown-deux==1.0.4
django-pagedown==0.0.4
django-toolbelt==0.0.1
gunicorn==17.5
markdown2==2.1.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
